SEQNG-162 Move SequenceDefaultToolbar to its own file #time 30m
